feat: read attributes from root node of XML documents
This commit modifies the Converter so that it now captures
attributes from the root nodes of XML documents. The logic change
for the Converter is relatively minor (read the attributes from
the XML reader before starting to recursively traverse the
document) so most of the changes in this commit are to add a few
variations of test to verify the new behaviour.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>